+++ boot-core/src/daemon/mod.rs
@@ -6,3 +6,25 @@ pub mod querier;
 pub mod sender;
 pub mod state;
 pub mod tx_resp;
+
+pub(crate) mod cosmos_modules {
+    pub use cosmrs::proto::cosmos::auth::v1beta1 as auth;
+    pub use cosmrs::proto::cosmos::authz::v1beta1 as authz;
+    pub use cosmrs::proto::cosmos::bank::v1beta1 as bank;
+    pub use cosmrs::proto::cosmos::base::abci::v1beta1 as abci;
+    pub use cosmrs::proto::cosmos::base::tendermint::v1beta1 as tendermint;
+    pub use cosmrs::proto::cosmos::base::v1beta1 as base;
+    pub use cosmrs::proto::cosmos::crisis::v1beta1 as crisis;
+    pub use cosmrs::proto::cosmos::distribution::v1beta1 as distribution;
+    pub use cosmrs::proto::cosmos::evidence::v1beta1 as evidence;
+    pub use cosmrs::proto::cosmos::feegrant::v1beta1 as feegrant;
+    pub use cosmrs::proto::cosmos::gov::v1beta1 as gov;
+    pub use cosmrs::proto::cosmos::mint::v1beta1 as mint;
+    pub use cosmrs::proto::cosmos::params::v1beta1 as params;
+    pub use cosmrs::proto::cosmos::slashing::v1beta1 as slashing;
+    pub use cosmrs::proto::cosmos::staking::v1beta1 as staking;
+    pub use cosmrs::proto::cosmos::tx::v1beta1 as tx;
+    pub use cosmrs::proto::cosmos::vesting::v1beta1 as vesting;
+    pub use cosmrs::proto::cosmwasm::wasm::v1 as cosmwasm;
+    pub use cosmrs::proto::tendermint::abci as tendermint_abci;
+}