+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unpublished
 
+## v0.8.0
+
+- Redo crate API exposure
+- Remove prelude module
+
 ## v0.4.0
 
 - Expose DaemonQuerier.